% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_function.R
\name{avg_mass}
\alias{avg_mass}
\title{Average Caterpillar Mass}
\usage{
avg_mass(alive1, mass1, alive2, mass2, alive3, mass3)
}
\arguments{
\item{alive1}{The number of caterpillars alive on day 1.}

\item{mass1}{The overall caterpillar mass on day 1.}

\item{alive2}{The number of caterpillars alive on day 2.}

\item{mass2}{The overall caterpillar mass on day 2.}

\item{alive3}{The number of caterpillars alive on day 3.}

\item{mass3}{The overall caterpillar mass on day 3.}
}
\description{
Find the average mass difference between two to three dates for a given container.
}
