% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_function.R
\name{avg_survival}
\alias{avg_survival}
\title{Caterpillar Survival}
\usage{
avg_survival(day1_alive, day2_alive, day3_alive)
}
\arguments{
\item{day1_alive}{The number of caterpillars alive before heat shock.}

\item{day2_alive}{The number of caterpillars alive after 1 heat shock.}

\item{day3_alive}{The number of caterpillars alive after 2 heat shocks.}
}
\description{
Calculate caterpillar survival after 1 or 2 heat shock exposures.
}
