% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development_function.R
\name{development_time}
\alias{development_time}
\title{Average Caterpillar Development Time}
\usage{
development_time(day1, day2, day3, day4)
}
\arguments{
\item{day1}{caterpillar hatch date}

\item{day2}{instar 2 date}

\item{day3}{instar 3 date}

\item{day4}{instar 4 date}
}
\description{
Calculate the average development time for a caterpillar between different instars.
}
